Fix incorrect number of arguments to mq_open()
Related issue:
https://gitee.com/openharmony/communication_services_softbus_lite/issues/I2BBKX

According to man:
If O_CREAT is specified in oflag, then two additional arguments
must be supplied.

os_adapter.c is missing "mode" parameter in mq_open(). It will compile
for MUSL, which does not check the number of arguments passed through
variadic argument but it doesn't mean it works as designed.

Fix this issue as well as build for other libraries with built-in
checks by introducing missing "mode" argument.

File mode has been selected to 0600 (RW for owner) but is a subject
to change if needed.

Signed-off-by: Zbigniew Bodek <zbigniew.bodek@huawei.com>